swal updated, charts added, primeng removed, user and news routing added

+++ admin/src/app/admin-app/components/html-content/html-content-list.component.ts
@@ -1,3 +1,4 @@
+import Swal from 'sweetalert2';
 import { Component, OnInit } from '@angular/core';
 import { HtmlContentService } from "./html-content.service";
 import { HtmlContentModel, HtmlContentResponse } from "./html-content.model";
@@ -12,25 +13,21 @@ import { MatTableDataSource } from '@angular/material';
 })
 
 export class HtmlContentComponent implements OnInit {
-
     objHtmlTemplate: HtmlContentModel = new HtmlContentModel();
     objResponse: HtmlContentResponse = new HtmlContentResponse();
     contentId: string;
-    showForm: boolean = false;
     dataSource:any;
     displayedColumns = ['SN','Title', 'Active', 'Actions'];    
     /* Pagination */
     pageSizeOptions = [5, 10, 25, 50, 100];
-  perPage:number = 10;
-  currentPage:number = 1;
-  totalItems:number = 1;
-  bindSort:boolean = false;
-  preIndex:number = 1;
+    perPage:number = 10;
+    currentPage:number = 1;
+    totalItems:number = 1;
+    bindSort:boolean = false;
+    preIndex:number = 1;
     /* End Pagination */
 
     ngOnInit() {
-        this.perPage = 10;
-        this.currentPage = 1;
         this.getHtmlEditorList();
     }
 
@@ -44,85 +41,48 @@ export class HtmlContentComponent implements OnInit {
     }
 
     errorMessage(objResponse: any) {
-        swal("Alert !", objResponse.message, "info");
-
+       Swal("Alert !", objResponse.message, "info");
     }
 
     bindList(objRes: HtmlContentResponse) {
         this.objResponse = objRes;
         this.dataSource = new MatTableDataSource(this.objResponse.dataList);    
-        this.preIndex = (this.perPage * (this.currentPage - 1));
         this.totalItems = objRes.totalItems;
-        if (objRes.totalItems) {
-            if (!this.bindSort) {
-                this.bindSort = true;
-                this.sortTable();
-            }
-            else
-                jQuery("table").trigger("update", [true]);
-
-        }
-
-    }
-
-    sortTable() {
-        setTimeout(() => {
-            jQuery('.tablesorter').tablesorter({
-                headers: {
-                    2: { sorter: false },
-                    3: { sorter: false }
-                }
-            });
-        }, 50);
     }
 
     edit(id: string) {
         this.router.navigate(['/html/editor',id]);
-        // this.showForm = true;
-        // this.contentId = id;
     }
 
     addHtml() {
         this.router.navigate(['/html/editor']);
-        // this.showForm = true;
-        // this.contentId = null;
-    }
-
-    showList(args) {
-        if (!args) {
-            this.getHtmlEditorList(); // if not
-        }
-        this.showForm = false;
-        this.sortTable();
     }
 
     delete(id: string) {
-        swal({
+      Swal({
             title: "Are you sure?",
             text: "You will not be able to recover this File !",
             type: "warning",
             showCancelButton: true,
             confirmButtonColor: "#DD6B55",
             confirmButtonText: "Yes, delete it!",
-            closeOnConfirm: false
-        },
-            () => {
+        })
+        .then((result) => {
+            if(result.value){
                 let objTemp: HtmlContentModel = new HtmlContentModel();
                 objTemp._id = id;
                 objTemp.deleted = true;
                 this._objService.deleteHtmlEditor(objTemp)
                     .subscribe(res => {
                         this.getHtmlEditorList();
-                        swal("Deleted!", res.message, "success");
+                        Swal("Deleted!", res.message, "success");
                     },
                         error => {
-                            swal("Alert!", error.message, "info");
-
+                            Swal("Alert!", error.message, "info");
                         });
+                    }
             });
-
     }
-
 
     pageChanged(event) {
         this.perPage = event.pageSize;
